+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_DESCRIPTION_BEGIN
+### NEXT_VERSION_DESCRIPTION_END
 ## [1.5.0]() (20-03-2019)
 
 * Methods that created variables delegates in `DashboardBuilder` are deprecated now and will be deleted in next major